fix: change cpuUsage type from float to long in StatsMessageResponse
The cpuUsage field was changed from float to long to better represent the data and avoid potential precision issues. This ensures consistency and accuracy in reporting CPU usage statistics.

+++ src/main/java/mindustrytool/servermanager/messages/response/StatsMessageResponse.java
@@ -11,7 +11,7 @@ import lombok.experimental.Accessors;
 public class StatsMessageResponse {
     public long ramUsage = 0;
     public long totalRam = 0;
-    public float cpuUsage = 0;
+    public long cpuUsage = 0l;
     public long players = 0;
     public String mapName = "";
     public byte[] mapData;
